+++ src/state/onboarding-recommendations.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { resolveWorkspaceStateIdentity } from "../agents/workspace-state-store.js";
+import { resolveWorkspaceStateIdentity } from "../agents/workspace-state-identity.js";
 import { sha256Hex } from "../infra/crypto-digest.js";
 import {
   executeSqliteQuerySync,
